Created separate HTML files using "templateUrl" instead of including the HTML code inside the Typescript components. --> easier to read and maintain & navigation through Typescript file is easier
