+++ src/app/app-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { LoginComponent } from './login.component';
-import { ChatComponent } from './chat.component';
-import { ProfileComponent } from './profile.component';
-import { SettingsComponent } from './settings.component';
-import { DirectoryComponent } from './directory.component';
-import { buyPRNComponent } from './buyPRN.component';
+import { LoginComponent } from './login/login.component';
+import { ChatComponent } from './chat/chat.component';
+import { ProfileComponent } from './profile/profile.component';
+import { SettingsComponent } from './settings/settings.component';
+import { DirectoryComponent } from './directory/directory.component';
+import { buyPRNComponent } from './buyPRN/buyPRN.component';
 
 const appRoutes: Routes = [
   { path: 'chat/:id', component: ChatComponent },
