+++ src/app/app.module.ts
@@ -4,14 +4,14 @@ import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 import { ServiceWorkerModule } from '@angular/service-worker';
 
-import { AppComponent } from './app.component';
+import { AppComponent } from './app/app.component';
 import { AppRoutingModule } from './app-routing.module';
-import { ChatComponent }  from './chat.component';
-import { LoginComponent }  from './login.component';
-import { ProfileComponent }  from './profile.component';
-import { SettingsComponent }  from './settings.component';
-import { DirectoryComponent }  from './directory.component';
-import { buyPRNComponent }  from './buyPRN.component';
+import { ChatComponent }  from './chat/chat.component';
+import { LoginComponent }  from './login/login.component';
+import { ProfileComponent }  from './profile/profile.component';
+import { SettingsComponent }  from './settings/settings.component';
+import { DirectoryComponent }  from './directory/directory.component';
+import { buyPRNComponent }  from './buyPRN/buyPRN.component';
 import { LinkyModule } from 'angular-linky';
 
 import { PipeModule }    from './pipes.module';
