+++ src/app/app/app.component.html
@@ -1,0 +1,91 @@
+<div class="appShell">
+  <img class="fullScreenImage" id="fullScreenImage" (click)="hideFullScreenImage()">
+  <progress value='0' max='100' id='uploader' class="uploadProgress">0%</progress>
+
+  <div class="topNav" [class.topNavFullWidth]="router.url.startsWith('/chat/')">
+    <div class="navLeft">
+      <button class="iconBtn" (click)="router.navigate(['profile','all'])" aria-label="Go to profile">
+        <img src="./../assets/App icons/Perrinn_02.png" class="brandIcon">
+      </button>
+      <button class="iconBtn" (click)="router.navigate(['directory'])" aria-label="Open directory">
+        <span class="material-icons">list</span>
+      </button>
+      <button class="iconBtn" (click)="openDiscover()" aria-label="Open discover">
+        <span class="material-icons-outlined">info</span>
+      </button>
+      <button class="iconBtn" (click)="toggleSocialLinksPopup()" aria-label="Open apps">
+        <span class="material-icons-outlined">apps</span>
+      </button>
+    </div>
+
+    <div class="navRight">
+      <button *ngIf="!UI.currentUser" class="buttonSecondary" (click)="router.navigate(['login'])" [disabled]='this.router.url.startsWith("/login")'>
+        Login
+      </button>
+
+      <div *ngIf="UI.currentUser" class="userPill" (click)="router.navigate(['profile',UI.currentUser])">
+        <img *ngIf="UI.currentUserLastMessageObj" [src]="UI.currentUserLastMessageObj?.imageUrlThumbUser" (error)="UI.handleUserImageError($event, UI.currentUserLastMessageObj)" class="avatar">
+        <span class="balance">{{UI.convertAndFormatPRNToPRNCurrency(null,UI.currentUserLastMessageObj?.wallet?.balance||0)}}</span>
+        <span *ngIf="UI.isCurrentUserMember" class="material-icons verified">check_circle</span>
+      </div>
+
+      <div class="devPillArea" *ngIf="UI.profileSimulatorNonMember || UI.profileSimulatorLoggedOut || UI.isDev || UI.revolutMode !== 'prod'">
+        <div class="devPill devPillActive devPillToggle" *ngIf="UI.profileSimulatorLoggedOut" (click)="UI.toggleprofileSimulatorLoggedOut()">Logged-out ON</div>
+        <div class="devPill devPillToggle" *ngIf="!UI.profileSimulatorLoggedOut && UI.isDev" (click)="UI.toggleprofileSimulatorLoggedOut()">Logged-out OFF</div>
+        <div class="devPill devPillActive devPillToggle" *ngIf="UI.profileSimulatorNonMember" (click)="UI.toggleprofileSimulatorNonMember()">Non-member ON</div>
+        <div class="devPill devPillToggle" *ngIf="!UI.profileSimulatorNonMember && UI.isDev" (click)="UI.toggleprofileSimulatorNonMember()">Non-member OFF</div>
+        <span *ngIf="UI.isDev" class="devPill">DEV</span>
+        <span *ngIf="UI.revolutMode=='sandbox'" class="devPill">sandbox</span>
+      </div>
+
+      <button class="buttonPrimary buyBtn" (click)="router.navigate(['buyPRN',''])" [disabled]='this.router.url.startsWith("/buyPRN")'>
+        Buy PRN
+      </button>
+    </div>
+  </div>
+
+  <div id='main_container'>
+    <div class="sideProfile" *ngIf="router.url.startsWith('/chat/')">
+      <profile [sidePanelScope]="'all'"></profile>
+    </div>
+    <div class="side-divider" [class.dragging]="isDragging" *ngIf="router.url.startsWith('/chat/')" (mousedown)="startDragging($event)">
+      <div class="side-divider-line"></div>
+    </div>
+    <div id='secondary_container' class="contentWrap" [class.contentWrapFullWidth]="router.url.startsWith('/chat/')">
+      <div class="contentCard">
+        <router-outlet></router-outlet>
+      </div>
+    </div>
+  </div>
+
+  <div class="socialPopupBackdrop" *ngIf="showSocialLinksPopup" (click)="closeSocialLinksPopup()">
+    <div class="socialPopupCard" (click)="$event.stopPropagation()">
+      <div class="socialPopupHeader">
+        <span class="socialPopupTitle">Apps</span>
+      </div>
+      <div class="socialPopupGrid">
+        <button *ngIf="UI.PERRINNProfileLastMessageObj?.publicLink" class="buttonSecondary socialPopupBtn" (click)="openSocialLink(UI.PERRINNProfileLastMessageObj?.publicLink)">
+          <span class="material-icons-outlined">cloud</span>
+          Google Drive
+        </button>
+        <button class="buttonSecondary socialPopupBtn" (click)="openSocialLink('https://chat.whatsapp.com/CzUNIrzBBuiI6lOCnh9DRx')">
+          <span class="material-icons-outlined">chat</span>
+          WhatsApp
+        </button>
+        <button class="buttonSecondary socialPopupBtn" (click)="openSocialLink('https://www.youtube.com/@PERRINN424WeAreATeam')">
+          <span class="material-icons-outlined">smart_display</span>
+          YouTube
+        </button>
+        <button class="buttonSecondary socialPopupBtn" (click)="openSocialLink('https://www.linkedin.com/company/perrinn')">
+          <span class="material-icons-outlined">business</span>
+          LinkedIn
+        </button>
+        <button class="buttonSecondary socialPopupBtn" (click)="openSocialLink('https://github.com/PERRINN')">
+          <span class="material-icons-outlined">code</span>
+          GitHub
+        </button><p>detail-product works!</p>
+        <h1>{{id}}</h1>
+      </div>
+    </div>
+  </div>
+</div>
